pyproject.toml
[tool.poetry]
name = "job-logger"
version = "0.1.0"
description = ""
authors = ["tyndev <17089219+tyndev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
httpx = "^0.27.0"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"
openai = "^1.16.1"
instructor = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
